refactor: move 'use-geo-map*.ts' to /src/hooks/
